Rename 'downstream' to the specific task - 'classify'. Rename 'models' module to 'embed'
