+++ examples/kgcn/animal_trade/README.md
@@ -58,7 +58,7 @@ The [main](../../examples/kgcn/animal_trade/main.py) function will:
 
 - Search Grakn for the k-hop neighbours of the selected examples, and store information about them as arrays, demoted in the code as `raw_arrays`. This data is saved to file so that subsequent steps can be re-run without recomputing these data
 
-- Build the TensorFlow computation graph using `model.KGCN`, including a multi-class classification step and learning procedure defined by `downstream.SupervisedKGCNClassifier`
+- Build the TensorFlow computation graph using `model.KGCN`, including a multi-class classification step and learning procedure defined by `classify.SupervisedKGCNClassifier`
 
 - Feed the `raw_arrays` to the TensorFlow graph, and perform learning
