+++ kglib/kgcn/embed/aggregation.py
