+++ kglib/kgcn/learn/classify.py
@@ -17,17 +17,35 @@
 #  under the License.
 #
 
+#
+#  Licensed to the Apache Software Foundation (ASF) under one
+#  or more contributor license agreements.  See the NOTICE file
+#  distributed with this work for additional information
+#  regarding copyright ownership.  The ASF licenses this file
+#  to you under the Apache License, Version 2.0 (the
+#  "License"); you may not use this file except in compliance
+#  with the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing,
+#  software distributed under the License is distributed on an
+#  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+#  KIND, either express or implied.  See the License for the
+#  specific language governing permissions and limitations
+#  under the License.
+#
+
 import tensorflow as tf
 import tensorflow.contrib.layers as layers
 
-import kglib.kgcn.models.metrics as metrics
-import kglib.kgcn.models.model
-import kglib.kgcn.preprocess.preprocess as preprocess
+import kglib.kgcn.embed.metrics as metrics
+import kglib.kgcn.embed.model
 
 
 class SupervisedKGCNClassifier:
 
-    def __init__(self, kgcn: kglib.kgcn.models.model.KGCN, optimizer, num_classes, log_dir, max_training_steps=10000,
+    def __init__(self, kgcn: kglib.kgcn.embed.model.KGCN, optimizer, num_classes, log_dir, max_training_steps=10000,
                  regularisation_weight=0.0, classification_dropout_keep_prob=0.7, use_bias=True,
                  classification_activation=lambda x: x, classification_regularizer=layers.l2_regularizer(scale=0.1),
                  classification_kernel_initializer=tf.contrib.layers.xavier_initializer()):
